fix: updated percentageSynced, name and updated version

+++ src/main/java/org/lamisplus/modules/central/domain/dto/SyncHistoryResponse.java
@@ -19,6 +19,6 @@ public class SyncHistoryResponse {
     private String facilityName;
     private Object errorLog;
     private UUID uuid;
-    private int percentageSynced;
+    private float percentageSynced;
     private String genKey;
 }
